fix: correct output read lengths from the simulator
The sequence simulator used to generate fragments should always output
the specified number of bases per read. Determine the length first and
then determine the offset of insertion and the fragment in the cDNA
sequence.

Set the insert length to a uniform distribution between 20 and 40. Too
long "insertion" should be mapped to the corresponding transcripts. Too
short insertion will not be mapped at all by any aligner.